#248: skipped some steps for redroid

+++ files/android.sh
@@ -33,11 +33,14 @@ else
   export AUTOMATION_NAME='uiautomator2'
 fi
 
-# uninstall appium specific applications
-adb uninstall io.appium.uiautomator2.server.test
-adb uninstall io.appium.uiautomator2.server
-adb uninstall io.appium.settings
-adb uninstall io.appium.unlock
+# there is no sense to clean something for scalable one-time redroid container
+if [ "$ANDROID_DEVICE" != "device:5555" ]; then
+  # uninstall appium specific applications
+  adb uninstall io.appium.uiautomator2.server.test
+  adb uninstall io.appium.uiautomator2.server
+  adb uninstall io.appium.settings
+  adb uninstall io.appium.unlock
 
-#127: android: clear /sdcard/*.mp4
-adb shell "rm -rf /sdcard/*.mp4"
+  #127: android: clear /sdcard/*.mp4
+  adb shell "rm -rf /sdcard/*.mp4"
+fi
